Add WorkerInterceptorBase and improve WorkerInterceptor docs to help new interceptor users
